+++ tests/lambda_executor_test.py
@@ -22,11 +22,11 @@
 
 import json
 import os
-from unittest import result
 
 import botocore.exceptions
+import cloudpickle as pickle
 import pytest
-from mock import AsyncMock, MagicMock, call
+from mock import AsyncMock, MagicMock
 
 from covalent_awslambda_plugin import AWSLambdaExecutor
 
@@ -634,3 +634,32 @@ async def test_query_task_exception_exception_path(lambda_executor, mocker):
         app_log_mock.exception.assert_called_once_with(client_error_mock)
         open_mock.assert_called_once_with(os.path.join(workdir, exception_filename), "r")
         json_load_mock.assert_called_once_with(open_mock.return_value.__enter__.return_value)
+
+
+def test_pickle_func_sync(lambda_executor):
+    """Test the synchronous function pickling method."""
+
+    def test_func(x):
+        return x
+
+    lambda_executor._pickle_func_sync(test_func, "/tmp", "test.pkl", [1], {"x": 1})
+    with open("/tmp/test.pkl", "rb") as f:
+        func, args, kwargs = pickle.load(f)
+
+    assert func(1) == 1
+    assert args == [1]
+    assert kwargs == {"x": 1}
+
+
+@pytest.mark.asyncio
+async def test_pickle_func(lambda_executor, mocker):
+    """Test the asynchronous function pickling method."""
+    pickle_func_sync_mock = mocker.patch(
+        "covalent_awslambda_plugin.awslambda.AWSLambdaExecutor._pickle_func_sync"
+    )
+
+    def test_func(x):
+        return x
+
+    await lambda_executor._pickle_func(test_func, "/tmp", "test.pkl", [1], {"x": 1})
+    pickle_func_sync_mock.assert_called_once_with(test_func, "/tmp", "test.pkl", [1], {"x": 1})